requirements.txt
dj-database-url==0.5.0
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
Markdown==3.1.1
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
